+++ src/fighter/common/status.rs
@@ -1,36 +1,11 @@
-pub mod dash;
-mod run;
-mod run_brake;
-mod jump_squat;
-mod jump;
-mod tread_jump;
-// mod tread_fall;
-mod guard_on;
+pub mod movement;
 mod guard;
-mod guard_off;
-mod guard_damage;
-mod escape;
-mod escape_air;
-pub mod escape_air_slide;
-mod passive;
-pub mod attack;
-mod attack_100;
-mod attack_dash;
-mod attack_3;
-// mod attack_3_common;
-mod attack_4;
-mod attack_air;
-mod catch;
-mod capture_pulled;
+pub mod escape;
 mod damage;
-mod damage_air;
-mod damage_fall;
-mod damage_fly;
-mod damage_fly_roll;
-// mod damage_fly_reflect;
-mod catched_air_ganon;
-mod landing;
-mod pass;
+pub mod attack;
+mod catch;
+
+mod passive;
 mod cliff;
 mod appeal;
 mod rebirth;
@@ -38,39 +13,14 @@ mod sub_transitions;
 mod sub_fighter;
 
 pub fn install() {
-    dash::install();
-    run::install();
-    run_brake::install();
-    jump_squat::install();
-    jump::install();
-    tread_jump::install();
-    // tread_fall::install();
-    guard_on::install();
+    movement::install();
     guard::install();
-    guard_off::install();
-    guard_damage::install();
     escape::install();
-    escape_air::install();
-    escape_air_slide::install();
-    passive::install();
-    attack::install();
-    attack_100::install();
-    attack_dash::install();
-    attack_3::install();
-    // attack_3_common::install();
-    attack_4::install();
-    attack_air::install();
-    catch::install();
-    capture_pulled::install();
     damage::install();
-    damage_air::install();
-    damage_fall::install();
-    damage_fly::install();
-    damage_fly_roll::install();
-    // damage_fly_reflect::install();
-    catched_air_ganon::install();
-    landing::install();
-    pass::install();
+    attack::install();
+    catch::install();
+
+    passive::install();
     cliff::install();
     appeal::install();
     rebirth::install();
