+++ src/fighter/common/status/attack/attack_100.rs
