+++ src/fighter/common/status/attack/attack_3_common.rs
