+++ src/fighter/common/status/catch/capture_pulled.rs
