+++ src/fighter/common/status/catch/catched_air_ganon.rs
