+++ src/fighter/common/status/damage/damage_air.rs
