+++ src/fighter/common/status/movement/jump_squat.rs
