+++ src/fighter/demon/status/dash.rs
@@ -1,5 +1,5 @@
 use crate::imports::status_imports::*;
-use crate::fighter::common::status::dash::*;
+use crate::fighter::common::status::movement::dash::*;
 
 #[status_script(agent = "demon", status = FIGHTER_DOLLY_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
 unsafe fn demon_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
