+++ src/fighter/dolly/status/dash_back.rs
@@ -1,5 +1,5 @@
 use crate::imports::status_imports::*;
-use crate::fighter::common::status::dash::*;
+use crate::fighter::common::status::movement::dash::*;
 
 #[status_script(agent = "dolly", status = FIGHTER_DOLLY_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
 unsafe fn dolly_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
