+++ src/fighter/ken/status/dash_back.rs
@@ -1,5 +1,5 @@
 use crate::imports::status_imports::*;
-use crate::fighter::common::status::dash::*;
+use crate::fighter::common::status::movement::dash::*;
 
 #[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
 unsafe fn ken_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
