+++ src/fighter/lucario/status/escape_air_slide.rs
@@ -1,5 +1,5 @@
 use crate::imports::status_imports::*;
-use crate::fighter::common::status::escape_air_slide::*;
+use crate::fighter::common::status::escape::escape_air_slide::*;
 
 #[status_script(agent = "lucario", status = FIGHTER_STATUS_KIND_ESCAPE_AIR_SLIDE, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_END)]
 unsafe fn lucario_escape_air_slide_end(fighter: &mut L2CFighterCommon) -> L2CValue {
