+++ src/fighter/marth/status/helper.rs
@@ -1,5 +1,5 @@
 use crate::imports::status_imports::*;
-use crate::fighter::common::status::attack::*;
+use crate::fighter::common::status::attack::attack::*;
 
 pub unsafe extern "C" fn marth_stance_cancel_helper(fighter: &mut L2CFighterCommon) -> L2CValue {
     if VarModule::is_flag(fighter.battle_object, marth::instance::flag::IS_STANCE) {
