+++ src/fighter/ryu/helper.rs
@@ -8,7 +8,7 @@ use {
     },
     smash_script::*,
     wubor_utils::table_const::*,
-    super::super::common::status::attack::only_jabs
+    super::super::common::status::attack::attack::only_jabs
 };
 
 pub unsafe extern "C" fn ryu_attack_main_uniq_chk(fighter: &mut L2CFighterCommon) -> L2CValue {
