+++ src/fighter/ryu/status.rs
@@ -1,20 +1,6 @@
 use {
-    smash::{
-        lua2cpp::L2CFighterCommon,
-        hash40,
-        phx::Hash40,
-        app::{lua_bind::*, *},
-        lib::{lua_const::*, L2CValue}
-    },
-    smash_script::*,
-    smashline::*,
-    custom_var::*,
-    wubor_utils::{
-        wua_bind::*,
-        vars::*,
-        table_const::*
-    },
-    super::super::common::status::dash::*,
+    crate::imports::status_imports::*,
+    super::super::common::status::movement::dash::*,
     super::helper::*
 };
